std.thread wouldn't compile

+++ internal/memset.d
@@ -25,34 +25,34 @@
 extern (C)
 {
     // Functions from the C library.
-    void *memcpy(void *, void *, uint);
+    void *memcpy(void *, void *, size_t);
 }
 
 extern (C):
 
-short *_memset16(short *p, short value, int count)
+short *_memset16(short *p, short value, size_t count)
 {
     short *pstart = p;
     short *ptop;
 
     for (ptop = &p[count]; p < ptop; p++)
-	*p = value;
+        *p = value;
     return pstart;
 }
 
-int *_memset32(int *p, int value, int count)
+int *_memset32(int *p, int value, size_t count)
 {
 version (X86)
 {
     asm
     {
-	mov	EDI,p		;
-	mov	EAX,value	;
-	mov	ECX,count	;
-	mov	EDX,EDI		;
-	rep			;
-	stosd			;
-	mov	EAX,EDX		;
+        mov     EDI,p           ;
+        mov     EAX,value       ;
+        mov     ECX,count       ;
+        mov     EDX,EDI         ;
+        rep                     ;
+        stosd                   ;
+        mov     EAX,EDX         ;
     }
 }
 else
@@ -61,59 +61,59 @@ else
     int *ptop;
 
     for (ptop = &p[count]; p < ptop; p++)
-	*p = value;
+        *p = value;
     return pstart;
 }
 }
 
-long *_memset64(long *p, long value, int count)
+long *_memset64(long *p, long value, size_t count)
 {
     long *pstart = p;
     long *ptop;
 
     for (ptop = &p[count]; p < ptop; p++)
-	*p = value;
+        *p = value;
     return pstart;
 }
 
-cdouble *_memset128(cdouble *p, cdouble value, int count)
+cdouble *_memset128(cdouble *p, cdouble value, size_t count)
 {
     cdouble *pstart = p;
     cdouble *ptop;
 
     for (ptop = &p[count]; p < ptop; p++)
-	*p = value;
+        *p = value;
     return pstart;
 }
 
-real *_memset80(real *p, real value, int count)
+real *_memset80(real *p, real value, size_t count)
 {
     real *pstart = p;
     real *ptop;
 
     for (ptop = &p[count]; p < ptop; p++)
-	*p = value;
+        *p = value;
     return pstart;
 }
 
-creal *_memset160(creal *p, creal value, int count)
+creal *_memset160(creal *p, creal value, size_t count)
 {
     creal *pstart = p;
     creal *ptop;
 
     for (ptop = &p[count]; p < ptop; p++)
-	*p = value;
+        *p = value;
     return pstart;
 }
 
-void *_memsetn(void *p, void *value, int count, int sizelem)
+void *_memsetn(void *p, void *value, int count, size_t sizelem)
 {   void *pstart = p;
     int i;
 
     for (i = 0; i < count; i++)
     {
-	memcpy(p, value, sizelem);
-	p = cast(void *)(cast(char *)p + sizelem);
+        memcpy(p, value, sizelem);
+        p = cast(void *)(cast(char *)p + sizelem);
     }
     return pstart;
 }
